TELEGRAM/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
numpy==1.25.2
paho-mqtt==1.6.1
requests==2.31.0
telepot==12.7
urllib3==2.0.4
yarl==1.9.2
